Fix spectrumCompMaxFrequency calculation to use float conversion for accurate bandwidth handling

+++ src/ScreenRadioBase.cpp
@@ -478,7 +478,7 @@ void ScreenRadioBase::setFftSamplingFrequencyAndSpektrumMaxDisplayFrequency() {
             bwStr = ::pSi4735Manager->getCurrentBandWidthLabelByIndex(Band::bandWidthSSB, config.data.bwIdxSSB);
         }
 
-        spectrumCompMaxFrequency = String(bwStr).toInt() * 1000; // Hangfrekvenciás sávszélesség KHz -> Hz konverzió
+        spectrumCompMaxFrequency = (uint16_t)(String(bwStr).toFloat() * 1000.0f); // Hangfrekvenciás sávszélesség KHz -> Hz konverzió (float konverzió)
         if (spectrumCompMaxFrequency < 1000) {
             spectrumCompMaxFrequency = 1000; // Minimum 1000 Hz (1 kHz) sávszélesség
         }
